Migrate Role to Collection.
It looks like the preferred way to have dependences on collections is to
be a collection yourself. As we're looking at developing a bunch of
roles and were trying to namespace them anyway this seems like a sane
thing to do.

I've asked for us to get the fossgalaxy namespace on galaxy so we can
publish these properly.